### 2019-08-17 #1657 @danielskatz
@sam-greenwood - you will need to change bibliography.bibtex to paper.bib
