### 2019-08-17 #1657 @danielskatz
👋 @sam-greenwood - I think merging https://github.com/sam-greenwood/taco_vis/pull/1 should allow the paper to compile - please enter ```@whedon generate pdf``` here as a new comment in this issue to try.
