### 2019-08-19 7614803
fixed cox et al (2013) citations
