### 2019-08-19 #1657 @sam-greenwood
Fixed the citations and reference
