### 2019-08-19 #1657 @sam-greenwood
> 👉 Check article proof 📄 👈

Almost, just need to fix the figures
