### 2019-08-31 #1657 @leouieda
:wave: Hi @banesullivan @malmans2 would you be interested in reviewing this submission for JOSS?
