### 2019-09-09 #1720 @whedon
[REVIEW]: Torsional Axisymmetric Core Oscillations Visualiser (TACO-VIS): A python module for animating torsional wave data for fluid planetary cores.
**Submitting author:** @sam-greenwood (<a href="http://orcid.org/0000-0001-9303-6229">Sam Greenwood</a>)
**Repository:** <a href="https://github.com/sam-greenwood/taco_vis" target ="_blank">https://github.com/sam-greenwood/taco_vis</a>
**Version:** v1.0
