### 2019-09-09 #1720 @leouieda
:wave: @malmans2 @banesullivan @sam-greenwood this is the review thread for the paper. All of our communications will happen here from now on.

Please feel free to ping me here (@leouieda) or email me privately if you have any questions/concerns.
