### 2019-09-30 #1720 @malmans2
Hi @sam-greenwood. I just started the review and I'm going to work on it this week. Sorry about the delay!
While reading the paper I noticed a couple of minor issues:

- [ ] Figure 1 is never mentioned.
- [ ] All references are missing a doi.

Other than that, I think it's well written and well structured.
I also have a suggestion about figure 2 and 3: Have you considered showing them in a single figure? They are based on the same data, so you could show them side-by-side and with just one colorbar.
