### 2019-09-30 sam-greenwood/taco_vis#2 @banesullivan
Proper name in LICENSE
https://github.com/sam-greenwood/taco_vis/blob/1a8ee79fb5475f84b50dcaa68ffac29b7bb22d80/LICENSE#L3

Could you please make sure the license properly list your name, not GitHub account?

per https://github.com/openjournals/joss-reviews/issues/1720
