### 2019-09-30 sam-greenwood/taco_vis#3 @banesullivan
Documentation and Examples
Is there work towards building a documentation site (possibly hosted on Read the Docs) that outlines the project, contains full API documentation, and provides reproducible, real-world examples?

per https://github.com/openjournals/joss-reviews/issues/1720
