### 2019-09-30 sam-greenwood/taco_vis#4 @banesullivan
Python Versions
It would be useful to have the compatible Python versions listed beyond just "Python 3" in the README. If all version of Python 3 are supported, the great! But I imagine they are not... it would be best to know which ones exactly.

I'd recommend committing to support 3.5+

In particular, it would be good to specify the `python_requires` argument to setup.py so that pip knows which versions are compatible.

per https://github.com/openjournals/joss-reviews/issues/1720
