### 2019-09-30 sam-greenwood/taco_vis#5 @banesullivan
list requirements in the setup.py
The requirements outlined in the README should be listed as the `install_requires` option in the `setup.py`

per https://github.com/openjournals/joss-reviews/issues/1720
