### 2019-09-30 sam-greenwood/taco_vis#6 @banesullivan
Add URL in the setup.py
add `url="https://github.com/sam-greenwood/taco_vis",` in the `setup.py`

per https://github.com/openjournals/joss-reviews/issues/1720
