### 2019-09-30 sam-greenwood/taco_vis#7 @banesullivan
Add proper classifiers to setup.py
add a `classifiers` section, perhaps something like:

```py
classifiers=(
        "Programming Language :: Python",
        "License :: OSI Approved :: MIT License",
        "Operating System :: OS Independent",
        'Topic :: Scientific/Engineering',
        'Topic :: Scientific/Engineering :: Visualization',
        'Intended Audience :: Science/Research',
        'Natural Language :: English',
    ),
```

per https://github.com/openjournals/joss-reviews/issues/1720
