### 2019-09-30 sam-greenwood/taco_vis#3 @malmans2
I was about to open the exact same issue!
