### 2019-09-30 sam-greenwood/taco_vis#8 @banesullivan
CI Testing
Where exactly are the tests?

We should get a CI service on Travis/Azure to run the tests across all supported Python versions (#4 is relevant here)

per https://github.com/openjournals/joss-reviews/issues/1720
