### 2019-09-30 #1720 @banesullivan
Hi @sam-greenwood, apologies for taking a while to get around to this! As a first pass, I just opened up quite a few issues in the repo that we can work together to resolve for this review.

I'll try to review the paper soon here as well!

Overall (from my first pass) it looks like the library is in need of:

- [ ] clean up around the install script (`setup.py` and repo meta info)
- [ ] clear tests and to validate it is working on a CI service like Travis - I can help set up those build scripts as we move along!
- [ ] a documentation website with examples and the full API docs should also be automatically generated - Read the Docs is great for this
