### 2019-09-30 sam-greenwood/taco_vis#10 @malmans2
PEP8 style conventions
I ran [pycodestyle](https://pypi.org/project/pycodestyle/) on your repository to check your code against the style conventions in [PEP 8](https://www.python.org/dev/peps/pep-0008/).
[Here](https://github.com/sam-greenwood/taco_vis/files/3673417/pycodestyle.log) is the output of ```pycodestyle . > pycodestyle.log```

Probably you can just manually fix everything, otherwise you could use an auto-formatter such as [black](https://black.readthedocs.io/en/stable/).

Also, you could connect your repository to a bot that automatically review new PR (e.g., [pep8speaks](https://pep8speaks.com/)).

per openjournals/joss-reviews#1720
