### 2019-09-30 sam-greenwood/taco_vis#8 @malmans2
I agree on this one too, and I suggest to check the code coverage as well (e.g., [codecov](https://codecov.io/))

Also, when I run `test.py` I get the same animation that is in the directory `test_files`, but the png images are different. The structure of the images is similar, but values are zero everywhere.
![](https://user-images.githubusercontent.com/22245117/65918886-d8092680-e3a8-11e9-86ed-661518f7ca5f.png)
