### 2019-09-30 sam-greenwood/taco_vis#11 @malmans2
Add community guidelines
Right now I don't see community guidelines. Please make clear whether you encourage external contributions or not, and create a section with instructions and best practices for contributors.

per openjournals/joss-reviews#1720
