### 2019-09-30 #1720 @malmans2
I agree with all bullet points by @banesullivan, especially on tests and documentation.
@sam-greenwood here are a few more:

- [ ] Follow the style guide for Python.
- [ ] Add community guidelines.
- [ ] Since your coauthors are not in the github contributors list, mention them in the documentation.

I haven't play much with the tools yet, I'm going to do it in the next couple of days!
