### 2019-10-02 #1720 @sam-greenwood
Thanks both for the feedback, I'll get to work making the changes and address any issues I've got questions about (which I'm sure there will be as I'm somewhat new to this!)
