### 2019-10-02 62b4594
added 'install' to installation instructions
