+++ setup.py
@@ -5,5 +5,14 @@ setup(name='taco_vis',
       description='Torsional Axisymmetric Core Oscillation Visualiser',
       author='Sam Greenwood',
       author_email='ee12sg@leeds.ac.uk',
-      py_modules=['taco_vis']
+      url="https://github.com/sam-greenwood/taco_vis",
+      py_modules=['taco_vis'],
+      classifiers=(
+        "Programming Language :: Python",
+        "License :: OSI Approved :: MIT License",
+        "Operating System :: OS Independent",
+        'Topic :: Scientific/Engineering',
+        'Topic :: Scientific/Engineering :: Visualization',
+        'Intended Audience :: Science/Research',
+        'Natural Language :: English')
      )
