### 2019-10-09 #1720 @sam-greenwood
> @banesullivan @malmans2 thank you for your reviews and opening issues on the project repo 👍
>
> @sam-greenwood please keep us informed of the progress and don't hesitate to ask questions or reach out if you need any help/clarification implementing the requested changes.

Will do, I've just been away on a brief holiday so haven't had a chance to continue work on this just yet but will try to address the majority of the issues this weekend.
