+++ taco_vis.py
@@ -6,28 +6,32 @@ from matplotlib.patches import Polygon
 from matplotlib.collections import PatchCollection
 import matplotlib.animation as anim
 
-########################################################################################
+###############################################################################
+
 
 class FLOW:
+    def __init__(self, data):
 
-    def __init__(self,data):
-
-        '''
+        """
         Initialise flow_class with a given data set.
 
         Parameters
         ----------
-        data: 3D or 2D numpy array where the values represent angular velocity. The shape of the
-        array should represent the number of data points in radius, theta and time respectively (for 3D)
-        or just radius and time (for 2D axisymmetric flows).
+        data: 3D or 2D numpy array where the values represent angular velocity.
+        The shape of the array should represent the number of data points in
+        radius, theta and time respectively (for 3D) or just radius and time
+        (for 2D axisymmetric flows).
 
-        e.g. if shape(data) = (5,10,20), then there are assumed 5 points in radius, 10 in theta
-        and 20 in time (for a 3D array) or if shape(data) = (5,10) then there are assumed 5 points in radius
-        and 20 in time (for a 2D array).
+        e.g. if shape(data) = (5,10,20), then there are assumed 5 points in
+        radius, 10 in theta and 20 in time (for a 3D array) or if
+        shape(data) = (5,10) then there are assumed 5 points in radius and 20
+        in time (for a 2D array).
 
-        For contour plotting 2D datasets, by default 50 points in theta are created. This can be changed by settings
-        the FLOW.th_resolution attribute to some other number after the class has been initialised if the user wishes.
-        '''
+        For contour plotting 2D datasets, by default 50 points in theta are
+        created. This can be changed by settings the FLOW.th_resolution
+        attribute to some other number after the class has been initialised if
+        the user wishes.
+        """
 
         a = data.shape
 
@@ -40,37 +44,38 @@ class FLOW:
             th = 0
             ti = a[1]
         else:
-            raise ValueError('Input data must be either a 2D or 3D array')
+            raise ValueError("Input data must be either a 2D or 3D array")
 
-
-        #Assume radius, theta and time arrays
+        # Assume radius, theta and time arrays
         self.data = data
-        self.radius = np.linspace(0,1,r)
-        self.theta = np.linspace(0,2*np.pi,th)
-        self.time = np.linspace(1,ti,ti)
+        self.radius = np.linspace(0, 1, r)
+        self.theta = np.linspace(0, 2 * np.pi, th)
+        self.time = np.linspace(1, ti, ti)
 
-        #Default Settings
+        # Default Settings
         self.speed = 1
 
-        self.colorbar_title = 'velocity'
-        self.title = 'time: %3.1f years'
+        self.colorbar_title = "velocity"
+        self.title = "time: %3.1f years"
 
-        self.movie_filename='output.mp4'
-        self.image_filename='output.png'
-        self.dpi=200
-        self.fps=24
+        self.movie_filename = "output.mp4"
+        self.image_filename = "output.png"
+        self.dpi = 200
+        self.fps = 24
 
         self.c_scale = np.max(np.abs(data[:]))
 
-        #Assume that if only 1 point in theta is given, flow is not theta dependant.
-        #Add in more theta points purely for contour plotting resolution
+        # Assume that if only 1 point in theta is given, flow is not theta
+        # dependant. Add in more theta points purely for contour plotting
+        # resolution
 
         if th == 0:
             self.th_resolution = 50
+
     #################################
 
-    #Define th_resolution as a property such that data is resized automatically if it changed
-    #from the default value.
+    # Define th_resolution as a property such that data is resized
+    # automatically if it changed from the default value.
[…]
