### 2019-10-13 328802a
Change style to match PEP8 (fixed line 753)
