### 2019-10-13 sam-greenwood/taco_vis#10 @sam-greenwood
Fixed the style, so now no output is given by running it through pycodestyle
