### 2019-10-13 89007ba
Update README.md
added bug reporting and contributors sections
