### 2019-10-13 sam-greenwood/taco_vis#11 @sam-greenwood
Added into the readme a section describing bug reporting and that external contributions are not currently being taken
