+++ setup.py
@@ -8,6 +8,8 @@ setup(
     author_email="ee12sg@leeds.ac.uk",
     url="https://github.com/sam-greenwood/taco_vis",
     py_modules=["taco_vis"],
+    python_requires='>=3',
+    install_requires=['numpy','matplotlib']
     classifiers=(
         "Programming Language :: Python",
         "License :: OSI Approved :: MIT License",
