### 2019-10-13 sam-greenwood/taco_vis#5 @sam-greenwood
Added install_requires to setup.py
