### 2019-10-13 sam-greenwood/taco_vis#8 @sam-greenwood
With regards to a CI service (something I am aware of but have never used before), I wasn't sure how that would work for testing a graphical output, rather than say computation of a number. What would you recommend would be the best way to do that?
