### 2019-10-13 sam-greenwood/taco_vis#8 @banesullivan
Have a [sphinx-gallery](https://sphinx-gallery.github.io) where examples are run and automatically generate figures which upload to a documentation website is a great way to test/verify that all is well
