### 2019-10-13 sam-greenwood/taco_vis#8 @banesullivan
Trying to implement an image comparison scheme wouldn't be worth your effort for a project this small
