### 2019-10-13 sam-greenwood/taco_vis#3 @sam-greenwood
Given that the user need only interact with the code with a few functions, I had originally intended and have tried to write the readme to cover the full usage of package. Do you feel this information would be better served on a separate documentation site or would it be fine in the readme (or both)?
