### 2019-10-13 sam-greenwood/taco_vis#3 @banesullivan
I think a separate site would be best... that way you can have the examples run and be tested automatically. I think this issue goes hand-in-hand with #8.

Examples and docs in READMEs often go untested and become outdated really quickly. Having something like a sphinx-gallery ensures that when a new user copy/pastes code from your example, it'll simply work.
