### 2019-10-30 #1720 @danielskatz
👋 @sam-greenwood - I'm just checking in on this as the associate-editor-in-chief on duty this week ...
