### 2019-10-30 #1720 @sam-greenwood
> 👋 @sam-greenwood - I'm just checking in on this as the associate-editor-in-chief on duty this week ...

Good thanks. I asked a question about CI testing to which I did get a prompt answer but didn't get a notification for it, so was waiting around the last 2 weeks for a response I'd already had, my bad. Just seen it now so will work on that this weekend.
