### 2019-11-03 #1720 @banesullivan
I finally got around to reviewing the paper! Overall I think it is well written and in line with most JOSS submissions, being a brief and high-level description of the software. I had the same concerns as @malmans2 in the [comment above](https://github.com/openjournals/joss-reviews/issues/1720#issuecomment-536698452) as well as a few other points:

- [ ] Figure 1 isn't mentioned (per @malmans2's comment), but also, was this figure made with the mentioned software? If not, you may consider removing it.

- [ ] Minor spelling/grammar: "python" should be "Python"

- It might be worth exploring combining figures 2, 3, and 4 as one figure with the three graphics in a single row (just a thought/suggestion)

- [ ] **State of the field requirement (per reviewer guidelines):** This part was a bit unclear to me from the paper. The following paragraph:

    > Although these torsional oscillations, described by the azimuthal component of velocity, uφ, can be visualised by static 2D (the cylindrical radius, s and time, t) plots (Cox, Livermore, & Mound, 2013; Gillet, Jault, Canet, & Fournier, 2010; Hide, Boggs, & Dickey, 2000; Teed, Jones, & Tobias, 2018), such diagrams do not communicate the geometry of the waves within the spherical core.

    appears to be addressing this, but it would be best to outline whether those references actually have software available or if they only describe the techniques to visualize these data. You then go on to mention that technique's limitations and how this software addresses those limitations, which is good, but I think it is worth mentioning any software out there that leverages those techniques.

- [ ] References. Make sure all have a DOI and:

    > In rapidly-rotating planets such as Earth, one important type of dynamical behaviour is longitudinal ‘torsional’ oscillations of concentric cylinders aligned with the rotation axis, each of which spans the height of the spherical core.

    This should be cited to point the reader to a paper/reference about this
