+++ examples/README.rst
@@ -1,0 +1,4 @@
+Example Gallery
+===============
+
+The following are reproducible examples showcasing the usage of ``taco_vis``
