+++ docs/source/index.rst
@@ -1,0 +1,28 @@
+.. taco_vis documentation master file, created by
+   sphinx-quickstart on Sun Nov  3 13:48:23 2019.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to taco_vis's documentation!
+====================================
+
+This needs to be filled out!
+
+
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+   :hidden:
+
+   self
+   examples/index
+
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
