### 2019-11-10 d7d687b
Added Travis CI file
