+++ .travis.yml
@@ -1,0 +1,15 @@
+language: python
+python:
+  - 3.5
+  - 3.6
+  - 3.7
+before_install:
+  - python --version
+  - pip install -U pip
+install:
+  - python setup.py install # install packages
+before_script: # configure a headless display to test plot generation
+- "export DISPLAY=:99.0"
+- "sh -e /etc/init.d/xvfb start"
+- sleep 3 # give xvfb some time to start
+script: python tests.py   # run tests
