+++ .travis.yml
@@ -8,8 +8,11 @@ before_install:
   - pip install -U pip
 install:
   - python setup.py install # install packages
-before_script: # configure a headless display to test plot generation
-- "export DISPLAY=:99.0"
-- "sh -e /etc/init.d/xvfb start"
-- sleep 3 # give xvfb some time to start
+dist: xenial
+services:
+  - xvfb
+# before_script: # configure a headless display to test plot generation
+# - "export DISPLAY=:99.0"
+# - "sh -e /etc/init.d/xvfb start"
+# - sleep 3 # give xvfb some time to start
 script: python tests.py   # run tests
