+++ setup.py
@@ -10,6 +10,7 @@ setup(
     py_modules=["taco_vis"],
     python_requires='>=3',
     install_requires=['numpy','matplotlib'],
+    setup_requires=['numpy','matplotlib'],
     classifiers=(
         "Programming Language :: Python",
         "License :: OSI Approved :: MIT License",
