### 2019-11-10 c09acbe
forced install of numpy and matplotlib
