+++ .travis.yml
@@ -6,6 +6,8 @@ python:
 before_install:
   - python --version
   - pip install -U pip
+  - pip install numpy
+  - pip install matplotlib
 install:
   - python setup.py install # install packages
 dist: xenial
