### 2019-11-10 2e983e4
added ffmpeg install
