+++ .travis.yml
@@ -8,6 +8,7 @@ before_install:
   - pip install -U pip
   - pip install numpy
   - pip install matplotlib
+  - pip install ffmpeg
 install:
   - python setup.py install # install packages
 dist: xenial
