### 2019-11-10 f752a89
added alternative ffmpeg install
