+++ .travis.yml
@@ -8,8 +8,9 @@ before_install:
   - pip install -U pip
   - pip install numpy
   - pip install matplotlib
-  - pip install ffmpeg
-  - which ffmpeg
+  - if [ "$TRAVIS_OS_NAME" == "linux" ]; then sudo add-apt-repository -y ppa:mc3man/trusty-media; fi
+  - if [ "$TRAVIS_OS_NAME" == "linux" ]; then sudo apt-get update; fi
+  - if [ "$TRAVIS_OS_NAME" == "linux" ]; then sudo apt-get install -y ffmpeg; fi
 install:
   - python setup.py install # install packages
 dist: xenial
