+++ .travis.yml
@@ -9,15 +9,10 @@ before_install:
   - pip install numpy
   - pip install matplotlib
   - if [ "$TRAVIS_OS_NAME" == "linux" ]; then sudo add-apt-repository -y ppa:mc3man/trusty-media; fi
-  - if [ "$TRAVIS_OS_NAME" == "linux" ]; then sudo apt-get update; fi
   - if [ "$TRAVIS_OS_NAME" == "linux" ]; then sudo apt-get install -y ffmpeg; fi
 install:
   - python setup.py install # install packages
 dist: xenial
 services:
   - xvfb
-# before_script: # configure a headless display to test plot generation
-# - "export DISPLAY=:99.0"
-# - "sh -e /etc/init.d/xvfb start"
-# - sleep 3 # give xvfb some time to start
 script: python tests.py   # run tests
