### 2019-11-10 04bb109
switched away from python setup.py install
