+++ .travis.yml
@@ -6,12 +6,10 @@ python:
 before_install:
   - python --version
   - pip install -U pip
-  - pip install numpy
-  - pip install matplotlib
   - if [ "$TRAVIS_OS_NAME" == "linux" ]; then sudo add-apt-repository -y ppa:mc3man/trusty-media; fi
   - if [ "$TRAVIS_OS_NAME" == "linux" ]; then sudo apt-get install -y ffmpeg; fi
 install:
-  - python setup.py install # install packages
+  - pip install . # install packages
 dist: xenial
 services:
   - xvfb
