### 2019-11-10 1d8b30c
test on all python > 3.0
