+++ .travis.yml
@@ -1,13 +1,18 @@
 language: python
 python:
+  - 3.0
+  - 3.1
+  - 3.2
+  - 3.3
+  - 3.4
   - 3.5
   - 3.6
   - 3.7
 before_install:
   - python --version
   - pip install -U pip
-  - if [ "$TRAVIS_OS_NAME" == "linux" ]; then sudo add-apt-repository -y ppa:mc3man/trusty-media; fi
-  - if [ "$TRAVIS_OS_NAME" == "linux" ]; then sudo apt-get install -y ffmpeg; fi
+  - sudo add-apt-repository -y ppa:mc3man/trusty-media
+  - sudo apt-get install -y ffmpeg
 install:
   - pip install . # install packages
 dist: xenial
