### 2019-11-10 sam-greenwood/taco_vis#8 @sam-greenwood
> I agree on this one too, and I suggest to check the code coverage as well (e.g., [codecov](https://codecov.io/))
>
> Also, when I run `test.py` I get the same animation that is in the directory `test_files`, but the png images are different. The structure of the images is similar, but values are zero everywhere.
> ![](https://user-images.githubusercontent.com/22245117/65918886-d8092680-e3a8-11e9-86ed-661518f7ca5f.png)

I have set up Travis-CI to test that the code can produce the plots, that the example data used has values above and below zero, and that it can save the plots. I test python 3.5 and up, with all of them passing the test. Could you let me know what version you were using to produce this plot with all zeros as I was unable to reproduce this.
