### 2019-11-10 sam-greenwood/taco_vis#4 @sam-greenwood
Added Travis-CI testing for confirmed 3.5+ support
