### 2019-11-10 sam-greenwood/taco_vis#3 @sam-greenwood
Ok, thanks for the work getting me started with the sphinx gallery, not something I've heard of before but looks useful!
