+++ .travis.yml
@@ -1,10 +1,5 @@
 language: python
 python:
-  - 3.0
-  - 3.1
-  - 3.2
-  - 3.3
-  - 3.4
   - 3.5
   - 3.6
   - 3.7
