### 2019-11-10 sam-greenwood/taco_vis#8 @malmans2
I still get the same problem for all figures produces by the test script (I tried from scratch).
This is what I did on my linux machine:
* Create a yml file with the packages to install:
```yml
name: taco_vis
channels:
  - conda-forge
  - defaults
dependencies:
  - python=3.7
  - numpy
  - matplotlib
  - ffmpeg
```
* Install and activate the new environment:
```
conda env create -f environment_taco_vis.yml
conda activate taco_vis
```
* Clone and install taco_vis in the environment:
```
git clone https://github.com/sam-greenwood/taco_vis.git
cd taco_vis/
python setup.py install
```
* Run tests:
```
python tests.py
```
I'm attaching a full list of packages installed: [pkgs_list.txt](https://github.com/sam-greenwood/taco_vis/files/3829081/pkgs_list.txt)
