### 2019-11-10 4e52900
Fixed some minor issues within paper
