### 2019-11-10 154768e
fixed issue in creating blank example files
