+++ tests.py
@@ -24,6 +24,7 @@ f = tv.FLOW(data)
 assert np.min(f.data) < 0, 'Data has no negative values'
 assert np.max(f.data) > 0, 'Data has no positive values'
 
+
 # Test animate contour plot_contours
 f.colorbar_title = "Non-dimensional\nvelocity"
 f.movie_filename = "test_contour.mp4"
@@ -32,12 +33,6 @@ f.plot_contours(animate=True, save=True)
 plt.close("all")
 assert os.path.isfile(f.movie_filename), 'File {} does not exist after saving'.format(f.movie_filename)
 
-
-# Create class
-f = tv.FLOW(data)
-
-# Resize number of theta points
-f.th_resolution = 10
 
 # Test contour plot_contours
 f.colorbar_title = "Non-dimensional\nvelocity"
@@ -48,16 +43,21 @@ plt.close("all")
 assert os.path.isfile(f.image_filename), 'File {} does not exist after saving'.format(f.movie_filename)
 
 
-f.image_filename = "test_cylinders.png"
-f.colorbar_title = "Non-dimensional\nvelocity"
-f.plot_cylinders(save=True, time_idx=14)
+#Create axisymmetric data
+data_axisym = flow_func(R, np.pi/2, T)
+
+f_axisym = tv.FLOW(data_axisym)
+
+f_axisym.image_filename = "test_cylinders.png"
+f_axisym.colorbar_title = "Non-dimensional\nvelocity"
+f_axisym.plot_cylinders(save=True, time_idx=14)
 # If plotting another image, close this animation figure first.
 plt.close("all")
-assert os.path.isfile(f.image_filename), 'File {} does not exist after saving'.format(f.movie_filename)
+assert os.path.isfile(f_axisym.image_filename), 'File {} does not exist after saving'.format(f_axisym.movie_filename)
 
 
-f.image_filename = "test_cylinders_3D.png"
-f.plot_cylinders_3D(save=True, time_idx=14)
+f_axisym.image_filename = "test_cylinders_3D.png"
+f_axisym.plot_cylinders_3D(save=True, time_idx=14)
 # If plotting another image, close this animation figure first.
 plt.close("all")
-assert os.path.isfile(f.image_filename), 'File {} does not exist after saving'.format(f.movie_filename)
+assert os.path.isfile(f_axisym.image_filename), 'File {} does not exist after saving'.format(f_axisym.movie_filename)
