### 2019-11-10 sam-greenwood/taco_vis#8 @sam-greenwood
Found the issue and fixed it (due to the nature of my example data. It was plotting the value of the data when theta=0, which in my example data set always gave 0). Thanks for finding that issue.
