+++ README.md
@@ -60,9 +60,6 @@ f.plot_cylinders(animate=False, save=False, time_idx=0)
 
 This produces a 2D plot of a slice through the equitorial plane with concentric circles representing the concentric cylinders (with as many cylinders as there are radial grid points in the data). A series of black dots are plotted on each cylinder and are advected to visualise the sense of rotation. The cylinders are coloured by the value of velocity at that time interval.
 
-<p align="center">
-  <img src="paper/images/example_cylinders.png" width="450" />
-</p>
 
 #### cylinders_3D
 
@@ -73,7 +70,7 @@ f.plot_cylinders_3D(animate=False, save=False, time_idx=0)
 The same as 'cylinders' but instead the plot is a 3D representation of the cylinders within a spherical core.
 
 <p align="center">
-  <img src="paper/images/example_cylinders_3D.png" width="450" />
+  <img src="paper/images/example_3D_and_2D.png" width="450" />
 </p>
 
 #### contours
