### 2019-11-25 #1720 @leouieda
:wave: hi @sam-greenwood @banesullivan @malmans2 just checking on the review. What is the progress on @banesullivan and @malmans2 questions?

> You then go on to mention that technique's limitations and how this software addresses those limitations, which is good, but I think it is worth mentioning any software out there that leverages those techniques.

I agree with @banesullivan on this. It would be good to mention (and cite) any software that does something similar, if it exists. If not, then a statement about the lack of software would also make a compelling case.
