### 2019-12-14 473cbc9
Add Travis build status to the README
