### 2019-12-30 ac8f226
Added images to the documentation
