+++ README.rst
@@ -1,49 +1,32 @@
-# Torsional Axisymmetric Core Oscillations Visualiser (TACO-VIS)
-A python module for animating torsional wave data for fluid planetary cores.
 
+Dependencies
+============
 
-TACO-VIS provides a simple set of python visualisation tools for fluid flow velocity data from fluid planetary interiors with a focus on animating torsional wave models. TACO-VIS is a lightweight module built only upon the common numpy/matplotlib python packages and is free to be used and modified as the user requires.
-
-Animations can be generated simply and quickly with minimal lines of code, e.g.:
-
-```py
-import numpy as np
-from taco_vis import FLOW
-
-# Import and generate some random data
-data = np.random.rand(10,10)
-
-f = FLOW(data)
-# Create instance of FLOW class and use it to plot an animation.
-f.plot_cylinders_3D(animate=True)
-```
-
-<p align="center">
-  <img src="paper/images/example_cylinders_3D.png" width="450" />
-</p>
-
-# Dependencies
 1. Python 3.5+
 2. numpy
 3. matplotlib
 4. ffmpeg framework (for saving animations as movie files)
 
-# Installation
+
+Installation
+============
 
 Python 3.5+, numpy and matplotlib are required for this module to produce the animations and ffmpeg to save them. Assuming you have at least python 3, numpy and matplotlib installed, clone/download this repository then run the setup file with `python setup.py install` from the command. This allows taco_vis to be imported from within the python environment, otherwise `taco_vis.py` must always be present in your working directory.
 
 ffmpeg can be obtained with instructions from https://ffmpeg.org
 
-# Usage
+Usage
+======
 
 Data must be angular velocity values in either a 2D or 3D numpy array. The dimensions of the array correspond to the number of points in radius/time (for 2D) or radius/azimuth/time (3D) repsectively. The data is also assumed to be on a regular grid with radius in the domain [0,1] and azimuth between [0, 2pi].
 
 Animations and figures are generated with the 'FLOW' class, which is initialised with the data array:
 
-```py
-from taco_vis import FLOW
-f = FLOW(data)
-```
+.. code-block::
+
+  from taco_vis import FLOW
+  f = FLOW(data)
+
 
 
 There are 3 types of plot available with the following methods: plot_contours, plot_cylinders and plot_cylinders_3D (detailed in full below). Animations or still images may be plotted and/or saved to disk for each of the style of plots provided. The same 3 keyword arguments are available for each:
@@ -52,59 +35,66 @@ There are 3 types of plot available with the following methods: plot_contours, p
 2. save (default = False). If True then the image (if animate is False) or animation (if animate is True) will not be shown but instead saved to the current working directory.
 3. time_idx (default = 0). If animate is False then this specifies the time index along the 3rd dimension of the data to be staically plotted.
 
-#### cylinders
 
-```py
-f.plot_cylinders(animate=False, save=False, time_idx=0)
-```
+cylinders
+---------
+
+.. code-block::
+
+  f.plot_cylinders(animate=False, save=False, time_idx=0)
+
 
 This produces a 2D plot of a slice through the equitorial plane with concentric circles representing the concentric cylinders (with as many cylinders as there are radial grid points in the data). A series of black dots are plotted on each cylinder and are advected to visualise the sense of rotation. The cylinders are coloured by the value of velocity at that time interval (right of figure below).
 
 
-#### cylinders_3D
+cylinders_3D
+-------------
 
-```py
-f.plot_cylinders_3D(animate=False, save=False, time_idx=0)
-```
+.. code-block::
+
+  f.plot_cylinders_3D(animate=False, save=False, time_idx=0)
+
 
[…]

[…]
