### 2019-12-30 sam-greenwood/taco_vis#3 @sam-greenwood
Ok created the documentation with sphinx.
