### 2019-12-30 #1720 @sam-greenwood
I think I have now addressed all of the issues. I have created the html documentation with sphinx and the comments by @banesullivan above:

1. Figure 1 is now cited at the end of the paragraph prior to it. The figure caption describes it as a cartoon if you feel this is sufficient?
2. Grammar issues changed
3. I have decided to combine 2 of the figure together but leave the last figure separate.
4. I have added in "To address this, we present the first visualisation tool specifically for these oscillations." to make this point clear.
5. All references now have a DOI
