### 2020-01-02 sam-greenwood/taco_vis#3 @banesullivan
@sam-greenwood, is the documentation automatically being built and deployed anywhere - such as on [readthedocs.org](https://readthedocs.org)?
