### 2020-01-06 sam-greenwood/taco_vis#3 @sam-greenwood
> @sam-greenwood, is the documentation automatically being built and deployed anywhere - such as on [readthedocs.org](https://readthedocs.org)?

I have tried to host it on readthedocs, but it's now error-ing whilst building the sphinx gallery, which I cannot reproduce locally. I have tried cloning the repo and using the same version of sphinx as readthedocs has used (1.8.5) which works on my local machine. Any suggestions of how I might resolve this?

https://readthedocs.org/projects/taco-vis/builds/10202526/
