### 2020-01-08 e9eb533
Fix spinx-gallery issue on read the docs
