+++ docs/source/conf.py
@@ -82,7 +82,7 @@ sphinx_gallery_conf = {
     # Sort gallery example by file name instead of number of lines (default)
     "within_subsection_order": FileNameSortKey,
     # directory where function granular galleries are stored
-    "backreferences_dir": False,
+    "backreferences_dir": None,
     # Modules for which function level galleries are created.  In
     "doc_module": "taco_vis",
     "image_scrapers": ('matplotlib'),
