### 2020-01-08 sam-greenwood/taco_vis#3 @banesullivan
#15 should fix that issue! Let me know when you get it deployed on readthedocs! 🚀
