### 2020-01-10 #1720 @malmans2
Hi @sam-greenwood and Happy New Year!
I'll take a look at everything in the next few days.
In the meantime, could you please re-compile the paper?
Also, it looks like readthedocs is not able to build the documentation: https://readthedocs.org/projects/taco-vis
I'm I looking at the wrong link or are you still working on it?
