+++ docs/source/conf.py
@@ -49,6 +49,8 @@ templates_path = ['_templates']
 # This pattern also affects html_static_path and html_extra_path.
 exclude_patterns = []
 
+#Set name of master rst file
+master_doc = 'index'
 
 # -- Options for HTML output -------------------------------------------------
 
