### 2020-01-10 sam-greenwood/taco_vis#3 @sam-greenwood
> #15 should fix that issue! Let me know when you get it deployed on readthedocs! 🚀

All fixed now, thanks!
