### 2020-01-10 #1720 @sam-greenwood
> Hi @sam-greenwood and Happy New Year!
> I'll take a look at everything in the next few days.
> In the meantime, could you please re-compile the paper?
> Also, it looks like readthedocs is not able to build the documentation: https://readthedocs.org/projects/taco-vis
> Am I looking at the wrong link or are you still working on it?

Hi @malmans2, happy new year to you too. Thanks for spotting that, I've fixed that now, forgot to commit an earlier change I made. I'll recompile the paper now.
