### 2020-01-10 #1720 @sam-greenwood
With the paper, is there a way to control the figure sizes (at the moment it's just expanding them to fit the page width)?
