### 2020-01-10 f996da3
Add documentation status badge
