### 2020-01-13 #1720 @leouieda
@sam-greenwood not really. What you can do for now is pad your figures with white space. Bear in mind that JOSS papers are not supposed to serve as documentation for the software. That's why we provide limited resources for styling and figures. Also note that the figure number is generated automatically, so you don't have to include "Figure 1" in your captions.
