### 2020-01-13 #1720 @leouieda
As @banesullivan suggested:

> It might be worth exploring combining figures 2, 3, and 4 as one figure with the three graphics in a single row (just a thought/suggestion)
