### 2020-01-13 #1720 @leouieda
:wave: Hi @banesullivan @malmans2 just checking on the status of this review after the holiday. I noticed that there are many unchecked items in the Functionality and Documentation sections. Have these been addressed, raised as issues, or not reviewed yet?
