### 2020-01-13 94e6c82
Update paper.md
removed figure numbers
