+++ paper/paper.md
@@ -35,7 +35,7 @@ The dynamics of liquid planetary cores is fundamental to planetary-scale phenome
 In rapidly-rotating planets such as Earth, one important type of dynamical behaviour is longitudinal 'torsional' oscillations of concentric cylinders aligned with the rotation axis, each of which spans the height of the spherical core [@cox2013forward] (Figure 1).
 
 
-![Figure 1. Cartoon conceptually showing the differential rotation of fluid cylinders within a sphere.](images/cartoon_cylinders.png)
+![Cartoon conceptually showing the differential rotation of fluid cylinders within a sphere.](images/cartoon_cylinders.png)
 
 
 From an initial starting configuration, the azimuthal (longitudinal) velocity distribution of the cylinders propagates across the core owing to the restoring force of the magnetic field that threads through the cylinders. Because of this spatially dependant modulating effect from the magnetic field, observations of torsional oscillations can constrain the strength of the magnetic field hidden deep inside the core.
@@ -44,13 +44,13 @@ Although these torsional oscillations, described by the azimuthal component of v
 
 To address this, we present the first visualisation tool specifically for these oscillations. The Torsional Axisymmetric Core Oscillation VISualiser (TACO-VIS) is a simple open-source Python module which takes core flow velocity data for torsional waves and produces either static plots or animations. This data may be visualised as a series of rotating concentric cylinders in either a full 3D visualisation or in 2D as a slice through the equatorial plane (Figure 2).
 
-![Figure 2. 3D (left) and 2D (right) visualisation of data from Cox et al. [@cox2013forward] approximated by 15 cylinders. The azimuthal velocity scale shown is non-dimensional.](images/example_3D_and_2D.png)
+![3D (left) and 2D (right) visualisation of data from Cox et al. [@cox2013forward] approximated by 15 cylinders. The azimuthal velocity scale shown is non-dimensional.](images/example_3D_and_2D.png)
 
 To make the cylindrical movement visible, we add a dotted-texture to each cylinder, which is advected as a tracer where positive velocities are taken to be in the eastward (anti-clockwise) direction. The number of cylinders is dictated by the number of radial points in the data and is therefore user controlled.
 
 The 2D equatorial plane visualisation may alternatively plot the velocity data by contours rather than concentric cylinders (Figure 3), useful as this does not strictly assume axisymmetry in the data and so flows other than torsional oscillations may actually be represented with this.
 
-![Figure 4. Simple contour plot of example data that varies sinusoidally in radius and azimuth.](images/example_contour.png)
+![Simple contour plot of example data that varies sinusoidally in radius and azimuth.](images/example_contour.png)
 
 
 Examples of how to generate animations as well as full documentation are included within the repository for this code.
