### 2020-01-13 sam-greenwood/taco_vis#17 @malmans2
Thoughts about settings
Hi @sam-greenwood,
I've a couple of thoughts about the settings. If I'm understanding the code right, currently you describe them in the [Usage section](https://taco-vis.readthedocs.io/en/latest/README.html#usage), and the way to list them is to initialize the FLOW class, then use `__call__`.

- Would it make sense to describe the default settings [here](https://github.com/sam-greenwood/taco_vis/blob/a75fcafd84a01d04120f2c80944791ce228d6bb0/taco_vis.py#L19-L39), so they will be automatically included in the documentation?
- Would it make sense to allow the user to customize these settings since the very beginning, and only use the defaults if they've not been provided?
- Maybe naming [this section](https://taco-vis.readthedocs.io/en/latest/README.html#module-taco_vis) `API reference` rather than `Documentation` is better?

per openjournals/joss-reviews#1720
