### 2020-01-13 sam-greenwood/taco_vis#18 @malmans2
Minor suggestions for the documentation
A few minor suggestions @sam-greenwood:

- [x]  You have both README.md and README.rst. I think you can get rid of README.md at this point, so you just have to edit one.
- [x]  I think something is not right in [README.md](https://github.com/sam-greenwood/taco_vis/blob/master/README.md) (and in the figure shown [here](https://taco-vis.readthedocs.io/en/latest/index.html#torsional-axisymmetric-core-oscillations-visualiser-taco-vis)). At the very beginning you mention that the dataset is generated with `data = np.random.rand(10,10)`, but then there are negative values and the colorbar shows values above 1 and below -1. I would expect a plot similar to the plot you have [here](https://taco-vis.readthedocs.io/en/latest/examples/random.html#sphx-glr-examples-random-py).
- [x]  What about double backquotes for the installation command [here](https://taco-vis.readthedocs.io/en/latest/README.html#installation)? ``` ``python setup.py install`` ```
- [x]  Add an empty line before ``.. code-block:`` [here](https://taco-vis.readthedocs.io/en/latest/README.html#known-issues)
- [x]  I don't think you need both of these sections: [Examples](https://taco-vis.readthedocs.io/en/latest/README.html#examples) and  [Example gallery](https://taco-vis.readthedocs.io/en/latest/examples/index.html#example-gallery) . I would combine them.
- [x]  I would quickly explain how to test the package locally. Maybe in the [Installation](https://taco-vis.readthedocs.io/en/latest/README.html#installation) section?
- [x]  In the example that uses Cox et al., what about adding a doi/link to the paper?

per openjournals/joss-reviews#1720
