### 2020-01-13 #1720 @malmans2
@leouieda I've checked most of the items, and I opened two new issues about API and documentation fixes.
