### 2020-02-01 #1720 @sam-greenwood
> @leouieda I've checked most of the items, and I opened two new issues about API and documentation fixes.

Happy to make the necessary changes for these, I'll get round to this next week.
