### 2020-02-01 sam-greenwood/taco_vis#17 @malmans2
Hi @sam-greenwood,
I just discovered [.setdefault()](https://www.tutorialspoint.com/python/dictionary_setdefault.htm) which might be useful!
