### 2020-02-03 #1720 @banesullivan
@sam-greenwood, I have gone through your latest changes to the paper and it looks great to me! I have checked off most items in the review above, but I would still like to see the following issues addressed:

- https://github.com/sam-greenwood/taco_vis/issues/8
- https://github.com/sam-greenwood/taco_vis/issues/18
- https://github.com/sam-greenwood/taco_vis/pull/16
