### 2020-02-03 sam-greenwood/taco_vis#8 @banesullivan
@sam-greenwood, is this issue resolved now that you have an example gallery and tests passing on Travis?
