### 2020-02-05 #1720 @leouieda
@banesullivan @malmans2 thank you for the input!

@sam-greenwood please let us know when these issues have been addressed so that the reviewers can update the checklist.
