### 2020-02-23 sam-greenwood/taco_vis#8 @sam-greenwood
Hi @banesullivan, yes it's all fixed now. I'll close the issue.
