### 2020-02-23 sam-greenwood/taco_vis#18 @sam-greenwood
Thanks for the ideas, I've implemented all of your changes
