+++ README.rst
@@ -1,143 +1,0 @@
-
-Dependencies
-============
-
-1. Python 3.5+
-2. numpy
-3. matplotlib
-4. ffmpeg framework (for saving animations as movie files)
-
-
-Installation
-============
-
-Python 3.5+, numpy and matplotlib are required for this module to produce the animations and ffmpeg to save them. Assuming you have at least python 3, numpy and matplotlib installed, clone/download this repository then run the setup file with `python setup.py install` from the command. This allows taco_vis to be imported from within the python environment, otherwise `taco_vis.py` must always be present in your working directory.
-
-ffmpeg can be obtained with instructions from https://ffmpeg.org
-
-Usage
-======
-
-Data must be angular velocity values in either a 2D or 3D numpy array. The dimensions of the array correspond to the number of points in radius/time (for 2D) or radius/azimuth/time (3D) repsectively. The data is also assumed to be on a regular grid with radius in the domain [0,1] and azimuth between [0, 2pi].
-
-Animations and figures are generated with the 'FLOW' class, which is initialised with the data array:
-
-.. code-block::
-
-  from taco_vis import FLOW
-  f = FLOW(data)
-
-
-
-There are 3 types of plot available with the following methods: plot_contours, plot_cylinders and plot_cylinders_3D (detailed in full below). Animations or still images may be plotted and/or saved to disk for each of the style of plots provided. The same 3 keyword arguments are available for each:
-
-1. animate (default = False). If True then a matplotlib animation is created, iterating through the time axis of the data. If False then simply a static plot of one time interval of the data will be created.
-2. save (default = False). If True then the image (if animate is False) or animation (if animate is True) will not be shown but instead saved to the current working directory.
-3. time_idx (default = 0). If animate is False then this specifies the time index along the 3rd dimension of the data to be staically plotted.
-
-
-cylinders
----------
-
-.. code-block::
-
-  f.plot_cylinders(animate=False, save=False, time_idx=0)
-
-
-This produces a 2D plot of a slice through the equitorial plane with concentric circles representing the concentric cylinders (with as many cylinders as there are radial grid points in the data). A series of black dots are plotted on each cylinder and are advected to visualise the sense of rotation. The cylinders are coloured by the value of velocity at that time interval (right of figure below).
-
-
-cylinders_3D
--------------
-
-.. code-block::
-
-  f.plot_cylinders_3D(animate=False, save=False, time_idx=0)
-
-
-The same as 'cylinders' but instead the plot is a 3D representation of the cylinders within a spherical core (left of figure below).
-
-
-.. image:: ./paper/images/example_3D_and_2D.png
-   :width: 450pt
-
-
-contours
-----------
-
-.. code-block::
-
-  f.plot_contours(animate=False, save=False, time_idx=0)
-
-
-A filled contour plot of the data is produces, which does not strictly need to be axisymmetric and hence the data array may be 3D (radius, theta, time).
-
-.. image:: ./paper/images/example_contour.png
-   :width: 450pt
-
-
-Default settings for the appearance of the plots are attributes of the FLOW class. Current settings can be seen by the call method of the class:
-
-.. code-block::
-
-  >>> from taco_vis import FLOW
-  >>> f = FLOW(data)
-  >>> f()
-
-  CURRENT SETTINGS ---------------
-  speed : 1
-  colorbar_title : velocity
-  title : time: %3.1f years
-  movie_filename : output.mp4
-  image_filename : output.png
-  dpi : 200
-  fps : 24
-  c_scale : 0.98
-  --------------------------------
-
-
-Speed: (float) Factor applied to the movement of the texture of the cylinders and cylinders_3D animations. A reasonable value is such that FLOW.speed x velocity is approximately 1.
-
-colorbar_title: (string) The title for the colorbar.
-
[…]
