### 2020-02-23 sam-greenwood/taco_vis#17 @sam-greenwood
Thanks @malmans2. I've made changes per your suggestions for the 1st and 3rd bullet point. With regards to the 2nd, I think it's simplest for the defaults to be initialised with the class, then all the user need do is change them before calling any of the methods to plot the data. The documentation and the readme on the GitHub page shows how to do this. What do you think?
