### 2020-02-23 #1720 @sam-greenwood
Hi @leouieda, @banesullivan @malmans2 , I've made the changes per the reviews suggestions (I apologise for the delay, writing up a thesis has kept me quite busy!). The one suggestion I didn't change is explained [here](https://github.com/sam-greenwood/taco_vis/issues/17).
