### 2020-02-24 #1720 @banesullivan
@leouieda: To me (after the latest updates), this paper is ready for acceptance! 🎉

Great work, @sam-greenwood!
