### 2020-02-25 sam-greenwood/taco_vis#18 @malmans2
Is the front page shown in the documentation automatically generated based on [README.md](https://github.com/sam-greenwood/taco_vis/blob/master/README.md), or does it point to a separate file? I see a [README.rst](https://github.com/sam-greenwood/taco_vis/blob/master/docs/source/README.rst) in the doc folder. I'll leave it up to you whether to have just a single or multiple READMEs, but make sure they are consistent. For example, the first figure in README.md points the plot with negative values, but you use `np.random.rand` in the code to reproduce it.
Other than that, everything looks good!
