### 2020-02-25 #1720 @malmans2
@sam-greenwood  thanks for addressing my points. I only have two optional suggestions:
- I left a comment [here]( https://github.com/sam-greenwood/taco_vis/issues/18#issuecomment-590884472).
- I would add the documentation status badge, which I believe is the PR started by @banesullivan.

@leouieda I also think the paper is good to go! Good job @sam-greenwood!
