### 2020-02-26 sam-greenwood/taco_vis#18 @sam-greenwood
The data plotted are all in the positive colours on the colour scale (since it was generated with `np.random.rand` as you say). The code always produces a colour scale centered on 0. This decision was taken since with torsional waves you would always expect to see both negative and positive values.
