### 2020-02-26 sam-greenwood/taco_vis#18 @malmans2
This is the figure in the README. https://github.com/sam-greenwood/taco_vis/blob/master/paper/images/example_cylinders_3D.png
Are you sure these values are all positive? Maybe it's just my eyes, but I see light blue while zero is set at green. Or maybe I'm missing something.
