### 2020-04-03 0d0e199
Editorial fixes and suggestions to the JOSS paper
Made a few suggestions and corrections to the wording of the paper.
Included a suggestion for a new section heading.
