+++ paper/paper.md
@@ -26,39 +26,40 @@ bibliography: bibliography.bibtex
 
 # Summary
 
-TACO-VIS provides a simple set of Python visualisation tools for 2D flow velocity data from fluid planetary interiors. It is mainly intended for animating torsional wave models for publication and presentation purposes. TACO-VIS is a lightweight module built only upon the common numpy/matplotlib Python packages and is free to be used and modified as the user requires.
+TACO-VIS provides a simple set of Python visualisation tools for 2D flow velocity data from fluid planetary interiors. It is mainly intended for animating torsional wave models for publication and presentation purposes. TACO-VIS is a lightweight module built only upon the common numpy and matplotlib Python packages and is free to be used and modified as the user requires.
 
 # Background
 
 The dynamics of liquid planetary cores is fundamental to planetary-scale phenomena such as the generation of a magnetic field. Understanding the relevant processes is a complex task because the fluid motion includes a wide range of convective and oscillatory behaviour. Visualisation of models, based either on observational data or computer simulations, is a key part of both scientific exploration and communication particularly to non-specialist audiences.
 
-In rapidly-rotating planets such as Earth, one important type of dynamical behaviour is longitudinal 'torsional' oscillations of concentric cylinders aligned with the rotation axis, each of which spans the height of the spherical core [@cox2013forward] (Figure 1).
+In rapidly-rotating planets such as the Earth, one important type of dynamic behaviour is longitudinal "torsional" oscillations of concentric cylinders aligned with the rotation axis, each of which spans the height of the spherical core [@cox2013forward] (Figure 1).
 
 
 ![Cartoon conceptually showing the differential rotation of fluid cylinders within a sphere.](images/cartoon.png)
 
 
-From an initial starting configuration, the azimuthal (longitudinal) velocity distribution of the cylinders propagates across the core owing to the restoring force of the magnetic field that threads through the cylinders. Because of this spatially dependant modulating effect from the magnetic field, observations of torsional oscillations can constrain the strength of the magnetic field hidden deep inside the core.
+From an initial starting configuration, the azimuthal (longitudinal) velocity distribution of the cylinders propagates across the core due to the restoring force of the magnetic field that threads through the cylinders. Because of this spatially dependant modulating effect from the magnetic field, observations of torsional oscillations can constrain the strength of the magnetic field hidden deep inside the core.
 
 Although these torsional oscillations, described by the azimuthal component of velocity, $u_\phi$, can be visualised by standard 2D (the cylindrical radius, $s$ and time, $t$) contour plots [@gillet2010fast; @cox2013forward; @hide2000angular; @teed2018torsional], such diagrams do not communicate the geometry of the waves within the spherical core.
 
-To address this, we present the first visualisation tool specifically for these oscillations. The Torsional Axisymmetric Core Oscillation VISualiser (TACO-VIS) is a simple open-source Python module which takes core flow velocity data for torsional waves and produces either static plots or animations. This data may be visualised as a series of rotating concentric cylinders in either a full 3D visualisation (Figure 2a) or in 2D as a slice through the equatorial plane (Figure 2b).
+# The TACO-VIS project
+
+To address this issue, we present the first visualisation tool specifically designed for these oscillations. The Torsional Axisymmetric Core Oscillation VISualiser (TACO-VIS) is a simple open-source Python module which takes core flow velocity data for torsional waves and produces either static plots or animations. This data may be visualised as a series of rotating concentric cylinders in either a full 3D visualisation (Figure 2a) or in 2D as a slice through the equatorial plane (Figure 2b).
 
[…]
