### 2020-06-15 #1720 @Kevin-Mattheus-Moerman
@sam-greenwood your paper is ready for processing and acceptance in JOSS. Please complete the steps outline above :point_up: let us know if you have any questions.

> @sam-greenwood I made a few editorial suggestions to the paper. After those are integrated please let me know. Tthere a couple of things that you'll need to do after that and before acceptance:
>
>     * Double check authors and affiliations (including ORCIDs)
>
>     * Make a release of the software with the latest changes from the review and **post the version number here**. This is the version that will be used in the JOSS paper.
>
>     * Archive the release on Zenodo/figshare/etc and **post the DOI here**.
>
>     * Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
