### 2020-06-20 caeb4ed
updated tests for python 3.8
