+++ .travis.yml
@@ -1,8 +1,8 @@
 language: python
 python:
-  - 3.5
   - 3.6
   - 3.7
+  - 3.8
 before_install:
   - python --version
   - pip install -U pip
