### 2020-06-20 9cce561
Update README.md
changed for python 3.6
