### 2020-06-20 a77cb2b
changed for python 3.6+
