### 2020-06-20 8474e15
v1.1 for python 3.6 and up
