### 2020-06-20 #1720 @sam-greenwood
Hi all, apologies about the long delay, it's been a difficult time recently. Hope all is well with everybody given the unique circumstances. @leouieda @Kevin-Mattheus-Moerman @kyleniemeyer  I've addressed those 4 points now:

* Double check authors and affiliations (including ORCIDs)
All double checked. One affiliation was updated. (I'll re-generate the pdf below such that it is updated)

* Make a release of the software with the latest changes from the review
Release of v1.1 has been created

* Archive the release on Zenodo/figshare/etc
DOI for Zenodo is: 10.5281/zenodo.3902334

* Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

The paper in the archive has the correct details.
