### 2020-06-21 #1720 @whedon
:wave: @openjournals/joss-eics, this paper is ready to be accepted and published.
